Split glibc from the busybox package
This change allows us to further separate busybox, shrinking its size.
The "downside" is that it now needs to be static. We loose roughly one
megabyte by doing so, but the upside is cleaner separation.

The glibc image is additionally smaller than its previous counterpart,
since the default USE-flag set is smaller (effectively removing pam).

Finally, change the dependency of s6 from busybox to glibc.