Don't try to match portage snapshot and stage3 date anymore.
Rationale:

Ideally the portage snapshot used is from the same date as the current stage3
image, however as the gentoo mirrors only keep a backlog of snapshots worth a week
this really never worked as intended for 2/3 of the time. In the lights of the
recent glibc vuln. it's probably better to always the latest snapshot.