Fix new cmd allowing invalid namespace dir when creating images #131
Core's `detect_namespace()` ensures we are working with a valid Kubler
namespace before any command runs, but makes an exception for the `new`
command as the user might want to create a new namespace.

As the further args to `new` are not parsed yet at the time `detect_namespace()`
is running, the new command itself needs to ensure the current working dir is
valid for anything but creating a new namespace.